+++ src/main/cliapp/src/components/AppMenu.js
