+++ src/main/cliapp/src/containers/layout/Layout2.js
@@ -1,0 +1,5 @@
+
+
+export const Layout2 = (props) => {
+  return <div>{props.children}</div>
+}
