+++ src/main/cliapp/src/containers/layout/index.js
@@ -1,0 +1,4 @@
+import { Layout } from "./Layout";
+import { Layout2 } from "./Layout2";
+
+export { Layout, Layout2 };
